refactor(client): extract UpdatePage tabs into standalone components
