+++ client/src/pages/UpdatePage.tsx
@@ -15,27 +15,13 @@ import { toast } from 'react-hot-toast';
 import {
   Download,
   RefreshCw,
-  CheckCircle,
   AlertTriangle,
   History,
   Settings,
   Package,
-  GitBranch,
   GitCommit,
-  Clock,
   Loader2,
   RotateCcw,
-  Zap,
-  Sparkles,
-  Bug,
-  Wrench,
-  Cog,
-  BookOpen,
-  TestTube,
-  Paintbrush,
-  CircleDot,
-  FileText,
-  Tag,
 } from 'lucide-react';
 import {
   checkForUpdates,
@@ -51,17 +37,17 @@ import {
   getAllReleases,
   getVersionHistory,
   isUpdateInProgress,
-  getChannelInfo,
   type UpdateCheckResponse,
   type UpdateProgressResponse,
   type UpdateConfig,
-  type UpdateChannel,
   type ReleaseNotesResponse,
   type ReleaseInfo,
   type VersionHistoryResponse,
 } from '../api/updates';
 import { handleApiError } from '../lib/errorHandling';
-import UpdateProgress from '../components/updates/UpdateProgress';
+import UpdateOverviewTab from '../components/updates/UpdateOverviewTab';
+import UpdateHistoryTab from '../components/updates/UpdateHistoryTab';
+import UpdateSettingsTab from '../components/updates/UpdateSettingsTab';
 
 const VersionsTab = lazy(() => import('../components/updates/VersionsTab'));
 
@@ -79,29 +65,6 @@ const tabs: Tab[] = [
   { id: 'settings', label: 'tabs.settings', icon: <Settings className="h-4 w-4" /> },
   { id: 'versions', label: 'tabs.versions', icon: <GitCommit className="h-4 w-4" /> },
 ];
-
-const CATEGORY_ICONS: Record<string, React.ReactNode> = {
-  sparkles: <Sparkles className="h-4 w-4" />,
-  bug: <Bug className="h-4 w-4" />,
-  zap: <Zap className="h-4 w-4" />,
-  wrench: <Wrench className="h-4 w-4" />,
-  cog: <Cog className="h-4 w-4" />,
-  'book-open': <BookOpen className="h-4 w-4" />,
-  'test-tube': <TestTube className="h-4 w-4" />,
-  paintbrush: <Paintbrush className="h-4 w-4" />,
-  'circle-dot': <CircleDot className="h-4 w-4" />,
-};
-
-const CATEGORY_COLORS: Record<string, string> = {
-  Features: 'text-emerald-400',
-  'Bug Fixes': 'text-rose-400',
-  Performance: 'text-amber-400',
-  Refactoring: 'text-sky-400',
-  Maintenance: 'text-slate-400',
-  Documentation: 'text-violet-400',
-  Tests: 'text-cyan-400',
-  Other: 'text-slate-400',
-};
 
 export default function UpdatePage() {
   const { t } = useTranslation(['updates', 'common']);
@@ -149,7 +112,7 @@ export default function UpdatePage() {
       const result = await getCurrentUpdate();
       setCurrentUpdate(result);
       return result;
-    } catch (err) {
+    } catch {
       setCurrentUpdate(null);
       return null;
     }
@@ -404,570 +367,43 @@ export default function UpdatePage() {
 
       {/* Tab Content */}
       {activeTab === 'overview' && (
-        <div className="space-y-6">
-          {/* Current Update Progress */}
-          {currentUpdate && isUpdateInProgress(currentUpdate.status) && (
-            <UpdateProgress
-              progress={currentUpdate}
-              onRollback={() => setShowRollbackConfirm(true)}
-              rollbackLoading={rollbackLoading}
-              onCancel={handleCancel}
-              cancelLoading={cancelLoading}
-            />
-          )}
-
-          {/* Version Info Cards */}
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            {/* Current Version */}
-            <div className="bg-slate-800 rounded-lg p-5 border border-slate-700">
-              <div className="flex items-center gap-2 mb-4">
-                <Package className="h-5 w-5 text-slate-400" />
-                <h3 className="font-medium text-white">{t('version.current')}</h3>
-              </div>
-              {checkResult && (
-                <div className="space-y-3">
-                  <div className="text-3xl font-bold text-white">
-                    v{checkResult.current_version.version}
-                  </div>
-                  <div className="flex items-center gap-2 text-sm text-slate-400">
-                    <GitBranch className="h-4 w-4" />
-                    <span className="font-mono">
-                      {checkResult.current_version.commit_short}
-                    </span>
-                    {checkResult.current_version.tag && (
-                      <span className="px-2 py-0.5 bg-slate-700 rounded text-xs">
-                        {checkResult.current_version.tag}
-                      </span>
-                    )}
-                  </div>
-                  <div className="flex items-center gap-2 text-sm">
-                    {(() => {
-                      if (checkResult.current_version.is_dev_build) {
-                        return (
-                          <span className="px-2 py-0.5 bg-amber-500/20 text-amber-400 rounded text-xs font-medium">
-                            {t('version.devBuild')}
-                          </span>
-                        );
-                      }
-                      return (
-                        <span className="text-emerald-400">Stable</span>
-                      );
-                    })()}
-                  </div>
-                </div>
-              )}
-            </div>
-
-            {/* Available Update */}
-            <div
-              className={`bg-slate-800 rounded-lg p-5 border ${
-                checkResult?.update_available
-                  ? 'border-blue-500/50 bg-blue-500/5'
-                  : !checkResult?.update_available && checkResult?.dev_version_available
-                    ? 'border-amber-500/30 bg-amber-500/5'
-                    : 'border-slate-700'
-              }`}
-            >
-              <div className="flex items-center gap-2 mb-4">
-                {checkResult?.update_available ? (
-                  <Zap className="h-5 w-5 text-blue-400" />
-                ) : !checkResult?.update_available && checkResult?.dev_version_available ? (
-                  <GitBranch className="h-5 w-5 text-amber-400" />
-                ) : (
-                  <CheckCircle className="h-5 w-5 text-emerald-400" />
-                )}
-                <h3 className="font-medium text-white">
-                  {checkResult?.update_available ? t('version.available') : t('version.upToDate')}
-                </h3>
-              </div>
-              {checkResult?.update_available && checkResult.latest_version ? (
-                <div className="space-y-3">
-                  <div className="text-3xl font-bold text-blue-400">
-                    v{checkResult.latest_version.version}
-                  </div>
-                  <div className="flex items-center gap-2 text-sm text-slate-400">
-                    <GitBranch className="h-4 w-4" />
-                    <span className="font-mono">
-                      {checkResult.latest_version.commit_short}
-                    </span>
-                  </div>
-                  {checkResult.last_checked && (
-                    <div className="flex items-center gap-2 text-xs text-slate-500">
-                      <Clock className="h-3 w-3" />
-                      {t('version.lastChecked')} {new Date(checkResult.last_checked).toLocaleString()}
-                    </div>
-                  )}
-                </div>
-              ) : (
-                <div className="space-y-3">
-                  <p className="text-slate-400">
-                    {t('version.upToDateDesc')}
-                  </p>
-                  {checkResult?.dev_version_available && checkResult.dev_version && (
-                    <div className="pt-2 border-t border-slate-700/50 space-y-2">
-                      <div className="flex items-center gap-2 text-sm text-amber-400">
-                        <GitBranch className="h-3.5 w-3.5" />
-                        <span className="font-medium">{t('version.devVersionAvailable')}</span>
-                      </div>
-                      <div className="flex items-center gap-3 text-xs text-slate-400">
-                        <span className="font-mono">{checkResult.dev_version.commit_short}</span>
-                        <span>{t('version.devCommitsAhead', { count: checkResult.dev_commits_ahead ?? 0 })}</span>
-                      </div>
-                      {checkResult.dev_commits && checkResult.dev_commits.length > 0 && (
-                        <div className="space-y-1 max-h-48 overflow-y-auto">
-                          {checkResult.dev_commits.map((commit) => (
-                            <div key={commit.hash_short} className="flex items-start gap-2 text-xs">
-                              <span className="font-mono text-slate-500 shrink-0">{commit.hash_short}</span>
-                              <span className="text-slate-300 break-all">
-                                {commit.type && (
-                                  <span className={`font-medium ${
-                                    commit.type === 'feat' ? 'text-emerald-400' :
-                                    commit.type === 'fix' ? 'text-blue-400' :
-                                    'text-slate-400'
-                                  }`}>
-                                    {commit.type}{commit.scope ? `(${commit.scope})` : ''}:
-                                  </span>
-                                )}{' '}
-                                {commit.message.includes(':') ? commit.message.split(':').slice(1).join(':').trim() : commit.message}
-                              </span>
-                            </div>
-                          ))}
-                        </div>
-                      )}
-                      {/* Dev Install Button */}
-                      <div className="pt-2">
-                        {!showDevUpdateConfirm ? (
-                          <button
-                            onClick={() => setShowDevUpdateConfirm(true)}
-                            disabled={devUpdateLoading || !!currentUpdate}
-                            className="flex items-center gap-2 px-3 py-1.5 bg-amber-600 hover:bg-amber-700 disabled:bg-slate-700 disabled:text-slate-500 text-white rounded-lg transition-all touch-manipulation active:scale-95 text-xs font-medium"
-                          >
-                            <Download className="h-3.5 w-3.5" />
-                            {t('buttons.installDevVersion')}
-                          </button>
-                        ) : (
-                          <div className="space-y-2">
-                            <p className="text-xs text-amber-400/80">
-                              <AlertTriangle className="h-3 w-3 inline mr-1" />
-                              {t('version.devWarning')}
-                            </p>
-                            <div className="flex items-center gap-2">
-                              <button
-                                onClick={handleStartDevUpdate}
-                                disabled={devUpdateLoading}
-                                className="flex items-center gap-1.5 px-3 py-1.5 bg-amber-600 hover:bg-amber-700 text-white rounded text-xs transition-all touch-manipulation active:scale-95"
-                              >
-                                {devUpdateLoading ? (
-                                  <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                                ) : (
-                                  <Download className="h-3.5 w-3.5" />
-                                )}
-                                {t('buttons.confirmDevInstall')}
-                              </button>
-                              <button
-                                onClick={() => setShowDevUpdateConfirm(false)}
-                                className="px-3 py-1.5 bg-slate-600 hover:bg-slate-500 text-white rounded text-xs transition-all touch-manipulation active:scale-95"
-                              >
-                                {t('common:cancel')}
-                              </button>
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Blockers Warning */}
-          {checkResult?.blockers && checkResult.blockers.length > 0 && (
-            <div className="bg-amber-500/10 border border-amber-500/30 rounded-lg p-4">
-              <div className="flex items-start gap-3">
-                <AlertTriangle className="h-5 w-5 text-amber-400 mt-0.5" />
-                <div>
-                  <h4 className="font-medium text-amber-400">{t('blockers.title')}</h4>
-                  <ul className="mt-2 space-y-1 text-sm text-slate-300">
-                    {checkResult.blockers.map((blocker, i) => (
-                      <li key={i}>• {blocker}</li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {/* Release Notes */}
-          {releaseNotes && releaseNotes.categories.length > 0 && (
-            <div className="bg-slate-800 rounded-lg p-5 border border-slate-700">
-              <div className="flex items-center gap-3 mb-1">
-                <FileText className="h-5 w-5 text-blue-400" />
-                <h3 className="font-medium text-white">{t('releaseNotes.title')}</h3>
-                <span className="text-sm font-mono text-slate-400">v{releaseNotes.version}</span>
-              </div>
-              {releaseNotes.previous_version && (
-                <p className="text-sm text-slate-500 mb-4 ml-8">
-                  {t('releaseNotes.since', { version: releaseNotes.previous_version })}
-                </p>
-              )}
-              <div className="space-y-4 ml-8">
-                {releaseNotes.categories.map((category) => (
-                  <div key={category.name}>
-                    <div className="flex items-center gap-2 mb-2">
-                      <span className={CATEGORY_COLORS[category.name] || 'text-slate-400'}>
-                        {CATEGORY_ICONS[category.icon] || <CircleDot className="h-4 w-4" />}
-                      </span>
-                      <h4 className={`text-sm font-medium ${CATEGORY_COLORS[category.name] || 'text-slate-400'}`}>
-                        {category.name}
-                      </h4>
-                    </div>
-                    <ul className="space-y-1 text-sm text-slate-300 ml-6">
-                      {category.changes.map((change, j) => (
-                        <li key={j}>• {change}</li>
-                      ))}
-                    </ul>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* Changelog */}
-          {checkResult?.update_available && checkResult.changelog.length > 0 && (
-            <div className="bg-slate-800 rounded-lg p-5 border border-slate-700">
-              <h3 className="font-medium text-white mb-4">{t('changelog.title')}</h3>
-              <div className="space-y-4">
-                {checkResult.changelog.map((entry, i) => (
-                  <div key={i} className="border-l-2 border-blue-500/50 pl-4">
-                    <div className="flex items-center gap-2 mb-2">
-                      <span className="font-medium text-white">v{entry.version}</span>
-                      {entry.is_prerelease && (
-                        <span className="px-2 py-0.5 bg-amber-500/20 text-amber-400 text-xs rounded">
-                          {t('changelog.prerelease')}
-                        </span>
-                      )}
-                    </div>
-                    {entry.changes.length > 0 && (
-                      <ul className="space-y-1 text-sm text-slate-300">
-                        {entry.changes.map((change, j) => (
-                          <li key={j}>• {change}</li>
-                        ))}
-                      </ul>
-                    )}
-                    {entry.breaking_changes.length > 0 && (
-                      <div className="mt-2">
-                        <span className="text-rose-400 text-sm font-medium">
-                          {t('changelog.breakingChanges')}
-                        </span>
-                        <ul className="space-y-1 text-sm text-rose-300">
-                          {entry.breaking_changes.map((change, j) => (
-                            <li key={j}>• {change}</li>
-                          ))}
-                        </ul>
-                      </div>
-                    )}
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* Update Button */}
-          {checkResult?.update_available && (
-            <div className="flex justify-end gap-3">
-              {!showUpdateConfirm ? (
-                <button
-                  onClick={() => setShowUpdateConfirm(true)}
-                  disabled={!checkResult.can_update || updateLoading || !!currentUpdate}
-                  className="flex items-center gap-2 px-6 py-3 bg-blue-600 hover:bg-blue-700 disabled:bg-slate-700 disabled:text-slate-500 text-white rounded-lg transition-all touch-manipulation active:scale-95 font-medium"
-                >
-                  <Download className="h-5 w-5" />
-                  {t('buttons.updateTo', { version: checkResult.latest_version?.version })}
-                </button>
-              ) : (
-                <div className="flex items-center gap-3 p-3 bg-slate-700 rounded-lg">
-                  <span className="text-sm text-slate-300">{t('buttons.confirmUpdate')}</span>
-                  <button
-                    onClick={handleStartUpdate}
-                    disabled={updateLoading}
-                    className="px-4 py-1.5 bg-blue-600 hover:bg-blue-700 text-white rounded text-sm transition-all touch-manipulation active:scale-95"
-                  >
-                    {updateLoading ? (
-                      <Loader2 className="h-4 w-4 animate-spin" />
-                    ) : (
-                      t('buttons.yesUpdate')
-                    )}
-                  </button>
-                  <button
-                    onClick={() => setShowUpdateConfirm(false)}
-                    className="px-4 py-1.5 bg-slate-600 hover:bg-slate-500 text-white rounded text-sm transition-all touch-manipulation active:scale-95"
-                  >
-                    {t('common:cancel')}
-                  </button>
-                </div>
-              )}
-            </div>
-          )}
-        </div>
+        <UpdateOverviewTab
+          t={t}
+          checkResult={checkResult}
+          currentUpdate={currentUpdate}
+          releaseNotes={releaseNotes}
+          updateLoading={updateLoading}
+          rollbackLoading={rollbackLoading}
+          cancelLoading={cancelLoading}
+          devUpdateLoading={devUpdateLoading}
+          showUpdateConfirm={showUpdateConfirm}
+          showDevUpdateConfirm={showDevUpdateConfirm}
+          onSetShowUpdateConfirm={setShowUpdateConfirm}
+          onSetShowDevUpdateConfirm={setShowDevUpdateConfirm}
+          onSetShowRollbackConfirm={setShowRollbackConfirm}
+          onStartUpdate={handleStartUpdate}
+          onStartDevUpdate={handleStartDevUpdate}
+          onCancel={handleCancel}
+        />
       )}
 
       {activeTab === 'history' && (
-        <>
-        {/* All Releases Section */}
-        <div className="mt-6 bg-slate-800 rounded-lg border border-slate-700 overflow-hidden">
-          <div className="px-4 py-3 border-b border-slate-700 bg-slate-800/80">
-            <h3 className="font-medium text-white flex items-center gap-2">
-              <Tag className="h-4 w-4 text-slate-400" />
-              {t('releases.title')}
-            </h3>
-            <p className="text-xs text-slate-400 mt-1">{t('releases.description')}</p>
-          </div>
-          {releasesLoading ? (
-            <div className="flex items-center justify-center py-8">
-              <Loader2 className="h-6 w-6 animate-spin text-blue-500" />
-            </div>
-          ) : releases.length === 0 ? (
-            <div className="px-4 py-8 text-center text-slate-400">
-              {t('releases.noReleases')}
-            </div>
-          ) : (
-            <div className="divide-y divide-slate-700/50">
-              {releases.map((release) => (
-                <div
-                  key={release.tag}
-                  className="px-4 py-3 flex items-center justify-between hover:bg-slate-700/30 transition-colors"
-                >
-                  <div className="flex items-center gap-3">
-                    <span className="font-mono text-white">{release.tag}</span>
-                    <span className="font-mono text-xs text-slate-500">{release.commit_short}</span>
-                  </div>
-                  <div className="flex items-center gap-3">
-                    {release.date && (
-                      <span className="text-sm text-slate-400">
-                        {new Date(release.date).toLocaleDateString()}
-                      </span>
-                    )}
-                    <span
-                      className={`inline-flex items-center px-2 py-0.5 rounded text-xs font-medium ${
-                        release.is_prerelease
-                          ? 'bg-amber-500/20 text-amber-400'
-                          : 'bg-emerald-500/20 text-emerald-400'
-                      }`}
-                    >
-                      {release.is_prerelease ? t('releases.prerelease') : t('releases.stable')}
-                    </span>
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-
-        {/* Version History Section */}
-        <div className="mt-6 bg-slate-800 rounded-lg border border-slate-700 overflow-hidden">
-          <div className="px-4 py-3 border-b border-slate-700 bg-slate-800/80">
-            <h3 className="font-medium text-white flex items-center gap-2">
-              <GitBranch className="h-4 w-4 text-slate-400" />
-              {t('versionHistory.title')}
-            </h3>
-            <p className="text-xs text-slate-400 mt-1">{t('versionHistory.description')}</p>
-          </div>
-          {versionHistoryLoading ? (
-            <div className="flex items-center justify-center py-8">
-              <Loader2 className="h-6 w-6 animate-spin text-blue-500" />
-            </div>
-          ) : !versionHistory || versionHistory.versions.length === 0 ? (
-            <div className="px-4 py-8 text-center text-slate-400">
-              {t('versionHistory.noHistory')}
-            </div>
-          ) : (
-            <div className="divide-y divide-slate-700/50">
-              {versionHistory.versions.map((entry) => {
-                const isCurrent =
-                  entry.version === versionHistory.current_version &&
-                  entry.git_commit_short === versionHistory.current_commit;
-                return (
-                  <div
-                    key={entry.id}
-                    className="px-4 py-3 hover:bg-slate-700/30 transition-colors"
-                  >
-                    <div className="flex items-center justify-between">
-                      <div className="flex items-center gap-3">
-                        <span className="font-mono text-white font-medium">
-                          v{entry.version}
-                        </span>
-                        <span className="font-mono text-xs text-slate-500">
-                          {entry.git_commit_short}
-                        </span>
-                        {entry.git_branch && (
-                          <span className="inline-flex items-center gap-1 px-2 py-0.5 rounded text-xs bg-slate-700 text-slate-300">
-                            <GitBranch className="h-3 w-3" />
-                            {entry.git_branch}
-                          </span>
-                        )}
-                        {isCurrent && (
-                          <span className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-emerald-500/20 text-emerald-400">
-                            {t('versionHistory.current')}
-                          </span>
-                        )}
-                      </div>
-                      <div className="flex items-center gap-3 text-sm text-slate-400">
-                        <span title={t('versionHistory.timesStarted', { count: entry.times_started })}>
-                          {t('versionHistory.timesStarted', { count: entry.times_started })}
-                        </span>
-                      </div>
-                    </div>
-                    <div className="mt-1.5 flex items-center gap-4 text-xs text-slate-500">
-                      <span>
-                        {t('versionHistory.firstSeen')}: {new Date(entry.first_seen).toLocaleDateString()}
-                      </span>
-                      <span>
-                        {t('versionHistory.lastSeen')}: {new Date(entry.last_seen).toLocaleDateString()}
-                      </span>
-                      {entry.python_version && (
-                        <span>
-                          {t('versionHistory.python')} {entry.python_version}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                );
-              })}
-            </div>
-          )}
-        </div>
-        </>
+        <UpdateHistoryTab
+          t={t}
+          releases={releases}
+          releasesLoading={releasesLoading}
+          versionHistory={versionHistory}
+          versionHistoryLoading={versionHistoryLoading}
+        />
       )}
 
       {activeTab === 'settings' && config && (
-        <div className="bg-slate-800 rounded-lg p-5 border border-slate-700 space-y-6">
-          {/* Auto-Check */}
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-white">{t('settings.autoCheck')}</h3>
-              <p className="text-sm text-slate-400">
-                {t('settings.autoCheckDesc')}
-              </p>
-            </div>
-            <label className="relative inline-flex items-center cursor-pointer">
-              <input
-                type="checkbox"
-                checked={config.auto_check_enabled}
-                onChange={(e) =>
-                  handleConfigChange('auto_check_enabled', e.target.checked)
-                }
-                disabled={configLoading}
-                className="sr-only peer"
-              />
-              <div className="w-11 h-6 bg-slate-700 peer-focus:outline-none rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-blue-600"></div>
-            </label>
-          </div>
-
-          {/* Check Interval */}
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-white">{t('settings.checkInterval')}</h3>
-              <p className="text-sm text-slate-400">{t('settings.checkIntervalDesc')}</p>
-            </div>
-            <select
-              value={config.check_interval_hours}
-              onChange={(e) =>
-                handleConfigChange('check_interval_hours', parseInt(e.target.value))
-              }
-              disabled={configLoading || !config.auto_check_enabled}
-              className="bg-slate-700 border border-slate-600 text-white rounded px-3 py-2 text-sm"
-            >
-              <option value={6}>{t('settings.every6Hours')}</option>
-              <option value={12}>{t('settings.every12Hours')}</option>
-              <option value={24}>{t('settings.every24Hours')}</option>
-              <option value={48}>{t('settings.every2Days')}</option>
-              <option value={168}>{t('settings.everyWeek')}</option>
-            </select>
-          </div>
-
-          {/* Update Channel */}
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-white">{t('settings.channel')}</h3>
-              <p className="text-sm text-slate-400">
-                {getChannelInfo(config.channel as UpdateChannel).description}
-              </p>
-            </div>
-            <select
-              value={config.channel}
-              onChange={(e) =>
-                handleConfigChange('channel', e.target.value as UpdateChannel)
-              }
-              disabled={configLoading}
-              className="bg-slate-700 border border-slate-600 text-white rounded px-3 py-2 text-sm"
-            >
-              <option value="stable">{t('settings.stable')}</option>
-              <option value="unstable">{t('settings.unstable')}</option>
-            </select>
-          </div>
-
-          {/* Auto Backup */}
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-white">{t('settings.autoBackup')}</h3>
-              <p className="text-sm text-slate-400">
-                {t('settings.autoBackupDesc')}
-              </p>
-            </div>
-            <label className="relative inline-flex items-center cursor-pointer">
-              <input
-                type="checkbox"
-                checked={config.auto_backup_before_update}
-                onChange={(e) =>
-                  handleConfigChange('auto_backup_before_update', e.target.checked)
-                }
-                disabled={configLoading}
-                className="sr-only peer"
-              />
-              <div className="w-11 h-6 bg-slate-700 peer-focus:outline-none rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-blue-600"></div>
-            </label>
-          </div>
-
-          {/* Require Healthy Services */}
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-white">{t('settings.requireHealthy')}</h3>
-              <p className="text-sm text-slate-400">
-                {t('settings.requireHealthyDesc')}
-              </p>
-            </div>
-            <label className="relative inline-flex items-center cursor-pointer">
-              <input
-                type="checkbox"
-                checked={config.require_healthy_services}
-                onChange={(e) =>
-                  handleConfigChange('require_healthy_services', e.target.checked)
-                }
-                disabled={configLoading}
-                className="sr-only peer"
-              />
-              <div className="w-11 h-6 bg-slate-700 peer-focus:outline-none rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-blue-600"></div>
-            </label>
-          </div>
-
-          {/* Last Check Info */}
-          {config.last_check_at && (
-            <div className="pt-4 border-t border-slate-700">
-              <div className="flex items-center gap-2 text-sm text-slate-400">
-                <Clock className="h-4 w-4" />
-                <span>
-                  {t('version.lastChecked')} {new Date(config.last_check_at).toLocaleString()}
-                </span>
-                {config.last_available_version && (
-                  <span className="text-blue-400">
-                    (v{config.last_available_version} available)
-                  </span>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
+        <UpdateSettingsTab
+          t={t}
+          config={config}
+          configLoading={configLoading}
+          onConfigChange={handleConfigChange}
+        />
       )}
 
       {activeTab === 'versions' && (
